Optimization: use lazyNode to hold entries
Before, every entry was contained in a regular Html.Styled.div node.
This meant that, for operations that didn't affect the currently
displayed entries list (such as refreshing the page and not getting any
new entries), every single entry was rebuilt in the virtual DOM.

Now, the entries are only updated if the arguments to renderEntry
change.

Also, deleting the first entry meant rebuilding every following
entry in the virtual DOM. Now, entries are keyed, and they can be
reordered, deleted and inserted in the virtual DOM without rebuilding
them.

This required adding a key to each node. This key is directly used as
the HTML ID for the node, and must be unique.

The key is generated using the URL, title and parent feed of each entry:
a constraint on the table storing them guarantees that their
combination is unique. However, they can contain characterss that aren't
allowed in HTML IDs.

As a result, non-unique ids can be generated by accident if two entries
differ by the substitution of one or more 'j' with non alphanumeric
chars.
If non-unique ids are present in the entries, every entry will be
displayed on-screen but with no CSS applied to them.

I estimate that this will be exceedingly rare, and worth the trade-off
of much faster updates: there was already a small but noticeable
(~200-500 ms ?) delay when deleting entries, when the current view lists
~100 of them.

https://guide.elm-lang.org/optimization/lazy.html
https://guide.elm-lang.org/optimization/keyed.html
https://package.elm-lang.org/packages/rtfeldman/elm-css/latest/Html.Styled.Keyed