Behavior of the thread that refreshes feeds in the background
Before, there was a thread that refreshed every feed and updated the
database in the background. To make the app more reactive, I forked a
new instance of this thread on each client requests for updates.
Obviously, this meant that the user spamming the refresh button would
fork many threads, wasting resources to perform simultaneous updates.

I used an MVar to implement some basic communications between threads:
now, there is a single background thread at all times. This thread loops,
waiting for either a set amount of time or until the signal to pre-empt
the refresh is received through the MVar.