Reduced memory usage.
Currently I run MPFS on a free fly.io instance, with only 256mb of
storage.

Occasionally, the app would have spikes in memory usage, exceed the limit and crash.

To avoid that, I have:
* Set RTS flags to specify a maximum heap size.
* In the function that refreshes feeds, previously every entry of every
  feed was fetched, concatenated and then stored. Now, we're fetching
  and storing entries one feed at a time.

After these changes, the memory usage is more or less constant (it
fluctuates in a ~5mb interval).